chore: refactor button components structure
