+++ app/components/Controls/index.ts
@@ -1,7 +1,17 @@
 export { Field } from "./Field";
 export { Fieldset } from "./Fieldset";
-export { Button } from "./Button";
 export { Input } from "./Input";
-export { Radio } from "./Radio";
-export { IconButton } from "./IconButton";
+export { Label } from "./Label";
 export { Select } from "./Select";
+export { Radio } from "./Radio";
+export { Checkbox } from "./Checkbox";
+export { Switch } from "./Switch";
+
+export {
+  Button,
+  ButtonLink,
+  Icon,
+  IconButton,
+  IconButtonLink,
+  type LucideIconsType,
+} from "./Button/index";
