+++ app/components/DirectoryView/DirectoryView.tsx
@@ -4,8 +4,7 @@ import { ReactNode, useEffect } from "react";
 import { DirectoryViewGrid } from "./DirectoryViewGrid";
 import { DirectoryViewTable } from "./DirectoryViewTable";
 import { NodeInfoModal } from "./NodeInfoModal";
-import { ButtonLink } from "../Controls/Button";
-import { Icon } from "../Controls/IconButton";
+import { ButtonLink, Icon } from "../Controls";
 import { H1 } from "../Fonts";
 import { useDirectoryStore } from "./useDirectoryStore";
 import { Container } from "~/components/Container";
