+++ app/routes/objects.route.tsx
@@ -17,7 +17,7 @@ import { requestDurationMiddleware } from "~/.server/requestDurationMiddleware";
 import { LZWDecoder } from "~/components/.client/ImageViewer/state/lzwDecoder";
 import { CrumbsOptions, getCrumbs } from "~/components/Breadcrumbs/getCrumbs";
 import { ClientOnly } from "~/components/ClientOnly";
-import { Button } from "~/components/Controls/Button";
+import { Button } from "~/components/Controls";
 import { DataGrid } from "~/components/DataGrid/DataGrid";
 import {
   buildDirectoryTree,
@@ -66,7 +66,11 @@ export const handle = {
     const prefix = data?.bucketConfig?.prefix ?? "";
 
     // Calculate the relative path (path after the data connection prefix)
-    const normalizedPrefix = prefix.endsWith("/") ? prefix : prefix ? `${prefix}/` : "";
+    const normalizedPrefix = prefix.endsWith("/")
+      ? prefix
+      : prefix
+        ? `${prefix}/`
+        : "";
     const relativePath =
       normalizedPrefix && pathName.startsWith(normalizedPrefix)
         ? pathName.slice(normalizedPrefix.length)
@@ -85,7 +89,9 @@ export const handle = {
 
     // Display name: show bucket name, or bucket/lastPrefixSegment if prefix exists
     const prefixLastSegment = prefix.replace(/\/$/, "").split("/").pop();
-    const dataConnectionName = prefix ? `${bucketName}/${prefixLastSegment}` : bucketName;
+    const dataConnectionName = prefix
+      ? `${bucketName}/${prefixLastSegment}`
+      : bucketName;
 
     const options: CrumbsOptions = {
       dataConnectionName,
@@ -159,7 +165,7 @@ export const loader = async ({
         bucketName,
         objectsWithUrls,
         bucketConfig.provider,
-        prefix
+        prefix,
       );
 
       return {
